docker-requirements.txt
# GraphRAG Pipeline Dependencies
# Python 3.10

# Core ML/NLP
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector Search
faiss-cpu==1.7.4
numpy==1.24.3

# Graph Database
neo4j==5.14.1

# LLM API
together==0.2.10

# Data Processing
pandas==2.1.3
jsonlines==4.0.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
requests==2.31.0

# Optional: Streamlit (for Phase 4 UI)
streamlit==1.29.0
